fbsimctl improvements
